Company profile (#51)
* bank account

* massive changes

* massive changes

* Delete resources/src/stores/bankAccount.ts

* massive changes

* massive changes

* add logs

---------

+++ app/src/Services/CompanySetting/CompanySettingService.php
@@ -30,8 +30,11 @@ readonly class CompanySettingService implements CompanySettingServiceInterface
      */
     public function update(CompanySetting|Model $model, array $attributes): bool
     {
-        $this->mediaService->delete($model->getPath());
-        $path = $this->mediaService->save($model->getId(), Arr::get($attributes, CompanySetting::PATH_COLUMN));
+        $path = $model->getPath();
+        if (!is_null($uploadedFile = Arr::get($attributes, CompanySetting::PATH_COLUMN))) {
+            $this->mediaService->delete($model->getPath());
+            $path = $this->mediaService->save($model->getId(), $uploadedFile);
+        }
 
         return $this->companySettingRepository->update(
             $model->getId(),
@@ -65,5 +68,10 @@ readonly class CompanySettingService implements CompanySettingServiceInterface
         $this->mediaService->delete($model->getPath());
 
         return $this->companySettingRepository->delete($model->getId(), $columnName);
+    }
+
+    public function getFirst(): ?CompanySetting
+    {
+        return $this->companySettingRepository->getFirst();
     }
 }
